+++ app/routes/components/CopyUiProvider/changelog.codegen.ts
@@ -12,4 +12,6 @@ export const changelog = ''
   + '- 2025-11-08: Add simple MDX support for `CopyUiProvider`.\n'
   + '- 2025-11-15: Export the `ColorName` and `ColorNo` types. Add colors cyan, teal,\n'
   + '  lime and orange to the theme.\n'
+  + '- 2025-11-17: Export type `Theme`. And Enhance internal MdxProvider to use more\n'
+  + '  sub-components of `Typography`.\n'
   ;