+++ src/components/CopyUiProvider/CHANGELOG.md
@@ -8,3 +8,5 @@
 - 2025-11-08: Add simple MDX support for `CopyUiProvider`.
 - 2025-11-15: Export the `ColorName` and `ColorNo` types. Add colors cyan, teal,
   lime and orange to the theme.
+- 2025-11-17: Export type `Theme`. And Enhance internal MdxProvider to use more
+  sub-components of `Typography`.